requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==10.2.0
python-dotenv==1.0.1
stripe==7.11.0
django-allauth==0.61.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-rest-framework==0.1.0 